requirements.txt
numpy
pandas
opencv-python
streamlit
tenserflow
pillow
av
tqdm